Renamed the concept of a "member" in a user-defined data type to a "field": - Renamed enumerated value Operator.Member to Operator.Field. - Renamed Expression.IsValidMemberSubject to IsValidFieldSubject. - Renamed class UserDataTypeMember.cs to UserDataTypeField.cs. Updated the handling of DimStatements in Compiler.cs to make variables declared in a DIM SHARED statement global. Added method Mapper.MakeGlobalVariable. Reworked CodeModel.Statements.DimStatement to support representing REDIM statements: - Added virtual property StatementName and method RenderPreserveFlag. - Reworked RenderImplementation to take these into account. - Added subclass CodeModel.Statements.RedimStatement overriding the new members. - Added enumerated value StatementType.Redim. - Added enumerated values REDIM and PRESERVE to TokenType. - Updated BasicParser.cs to handle parsing REDIM statements. Added parameter requireSubscripts to ParseVariableDeclaration. - Added automated tests of REDIM statement parsing. Added part of the translation for DimStatements and RedimStatements to Compiler.cs: - Added statement type DimensionArrayStatement.cs. - Implemented translation in Compiler.TranslateStatement. - Left a nice chunky TODO. Implemented user data type variables & values: - Moved UserDataTypeValue.cs into ExecutionEngine/Execution/Variables and renamed it to UserDataTypeVariable.cs. Fleshed out its implementation as a subclass of Variable. - Updated Variable.Construct to handle user types using, in fact, exactly the suggested implementation from the TODO. - Added an overload of TypeRepository.ResolveType for resolving user data types. - Added class FieldAccessExpression.cs to ExecutionEngine/Compiled/Expressions. - Fixed the translation of IdentifierExpressions in Compiler.TranslateExpression to get the actual variable type rather than just infer what the type _would_ be for the identifier. Added method Mapper.GetVariableType. - Added a check to the translation of BinaryExpression for the Field operator, translating to FieldAccessExpression. Added exception template method CompilerExceptions.ElementNotDefined (why do they call it element here and field everywhere else?). Continued implementation of array variables: - Added classes ArraySubscriptExpressions.cs and ArraySubscriptsExpressions.cs to ExecutionEngine/Compiled, mirroring the ArraySubscripts and ArraySubscript classes but with evaluable expressions for the bounds rather than scalar integers. - Added type ArrayVariable.cs to have a dynamically-allocated Array be a variable's value. Added helper method DataType.MakeElementType. Added a static Uninitialized field to Array.cs along with a property IsUninitialized. - Implemented Variable.ConstructArray by instantiating ArrayVariable. - Changed the data types of LowerBound and UpperBound in ArraySubscript.cs from int to short. Added a debug helper ToString implementation to UserDataType.cs. Added an annotation to DataType.IsUserType so that if it returns true, the nullity analyzer can assume that the UserType member is not null. Added exception template method CompilerException.IllegalInSubFunctionOrDefFn.

+++ QBX/ExecutionEngine/Execution/ArraySubscript.cs
@@ -2,8 +2,8 @@
 
 public class ArraySubscript
 {
-	public int LowerBound;
-	public int UpperBound;
+	public short LowerBound;
+	public short UpperBound;
 
 	public int ElementCount => UpperBound - LowerBound + 1;
 }
